add size_t

+++ libft/ft_memchr.c
@@ -6,7 +6,7 @@
 /*   By: maguzman <maguzman@student.42.fr>         #+#  +:+       +#+         */
 /*                                               +#+#+#+#+#+   +#+            */
 /*   Created: 2026/05/07 19:42:07 by maguzman         #+#    #+#              */
-/*   Updated: 2026/05/07 20:43:56 by maguzman        ###   ########.fr        */
+/*   Updated: 2026/05/27 10:34:33 by maguzman        ###   ########.fr        */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,14 +15,14 @@
 void	*ft_memchr(const void *s, int c, size_t n)
 {
 	const unsigned char	*str;
-	long unsigned int	i;
+	size_t				i;
 
 	str = (const unsigned char *) s;
 	i = 0;
 	while (i < n)
 	{
 		if (str[i] == (unsigned char) c)
-			return ((void *) & str[i]);
+			return ((void *) &str[i]);
 		i++;
 	}
 	return (NULL);
